requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
Werkzeug==2.3.7
WTForms==3.0.1
SQLAlchemy==2.0.21
email-validator==2.0.0
python-slugify==8.0.1
Pillow==10.0.1
